requirements.txt
streamlit
google-generativeai
python-dotenv
twython
ConfigParser
tweepy
streamlit 
requests
python-facebook-api
praw==7.4.0
python-dotenv==0.19.1
Pillow==8.3.2
requests==2.26.0
facebook-sdk==3.1.0